+++ lib/material_drum_picker.dart
@@ -6,6 +6,8 @@ library;
 // Models
 export 'src/models/drum_calendar_type.dart';
 export 'src/models/drum_column_order.dart';
+export 'src/models/drum_date_format.dart';
+export 'src/models/drum_month_format.dart';
 export 'src/models/drum_picker_labels.dart';
 export 'src/models/drum_picker_mode.dart';
 export 'src/models/drum_picker_value.dart';